Create Value 2.sql
IN Microsoft Access you may get "Microsoft Access can't append all the records in the append query" but ignore it and press yes to append the record into the table. Make sure the table isn't open as to not get an error. Since this is the second record, the [example] helps add another record.